feat(Refactoring): Change folder name to apparalls
