+++ src/Components/Category/AllCategory.jsx
@@ -1,12 +1,12 @@
-import Croptop from '../../assets/clothings/Croptop.jpg'
-import Fastrackwatch from '../../assets/clothings/Fastrackwatch.webp'
-import Menshirt from '../../assets/clothings/Menshirt.webp'
-import Menshoes from '../../assets/clothings/Menshoes.jpg'
-import Suit from '../../assets/clothings/Suit.jpg'
-import Suit2 from '../../assets/clothings/Suit2.jpg'
-import Tshirt from '../../assets/clothings/Tshirt.jpg'
-import Womenjeans from '../../assets/clothings/Womenjeans.jpg'
-import WomenShoes from '../../assets/clothings/WomenShoes.jpg'
+import Croptop from '../../assets/apparalls/Croptop.jpg'
+import Fastrackwatch from '../../assets/apparalls/Fastrackwatch.webp'
+import Menshirt from '../../assets/apparalls/Menshirt.webp'
+import Menshoes from '../../assets/apparalls/Menshoes.jpg'
+import Suit from '../../assets/apparalls/Suit.jpg'
+import Suit2 from '../../assets/apparalls/Suit2.jpg'
+import Tshirt from '../../assets/apparalls/Tshirt.jpg'
+import Womenjeans from '../../assets/apparalls/Womenjeans.jpg'
+import WomenShoes from '../../assets/apparalls/WomenShoes.jpg'
 import ItemComponent from './ItemComponent'
 import {Link} from 'react-router-dom'
 const AllCategory = () => {
